UserProfile controller & service, methods for getting a profile and updating it

+++ MireaHackBack/Repository/IUserProfileRepository.cs
@@ -5,8 +5,8 @@ namespace MireaHackBack.Repository;
 public interface IUserProfileRepository
 {
     public IQueryable<UserProfile> GetUsersProfiles();
-    public UserProfile? GetUserProfileById(int id);
-    public UserProfile? GetUserProfileByUserId(int userId);
+    public UserProfile? GetUserProfileById(long id);
+    public UserProfile? GetUserProfileByUserId(long userId);
     public UserProfile? GetUserProfileByUsername(string username);
     public UserProfile? GetUserProfileByEmail(string email);
     public bool CreateUserProfile(UserProfile user);
